Encountered CMP0004 error
Author: Stig-Ørjan Smelror <smelror@gmail.com>
Date: 2018-02-23

While compiling ecFlow with Python 3 support I encountered
a CMP0004 error:

Target "xxxx" links to item " -L/usr/lib64 -lpython3" which has
  leading or trailing whitespace.  This is now an error according to policy
  CMP0004.

The cause of this error is that $LIBPLUSED is not always set.
The fix is to move this to the second place.

Index: Python-3.7.4/Misc/python-config.sh.in
===================================================================

patch: python3-3.6.2-python3-config-LIBPLUSED-cmp0004-error.patch from upstream